+++ io_revolt/__init__.py
@@ -16,25 +16,55 @@ from bpy.app.handlers import persistent  # For the scene update handler
 from . import (
     common,
     layers,
-    panels,
+    operators,
+    texanim,
+    tools,
+)
+
+from .props import (
     props_mesh,
     props_obj,
     props_scene,
-    operators,
+)
+
+from .ui import (
+    menu_add,
+    headers,
+    faceprops,
+    instances,
+    light,
+    hull,
+    object,
+    scene,
+    vertex,
     texanim,
-    tools
+    helpers,
+    settings,
 )
 
 # Reloads potentially changed modules on reload (F8 in Blender)
 imp.reload(common)
 imp.reload(layers)
-imp.reload(panels)
 imp.reload(props_mesh)
 imp.reload(props_obj)
 imp.reload(props_scene)
 imp.reload(operators)
 imp.reload(texanim)
 imp.reload(tools)
+
+# Reloads ui
+imp.reload(menu_add)
+imp.reload(headers)
+imp.reload(faceprops)
+imp.reload(instances)
+imp.reload(light)
+imp.reload(hull)
+imp.reload(object)
+imp.reload(scene)
+imp.reload(vertex)
+imp.reload(texanim)
+imp.reload(helpers)
+imp.reload(settings)
 
 # Reloaded here because it's used in a class which is instanced here
 if "fin_in" in locals():
@@ -43,6 +73,8 @@ if "fin_out" in locals():
     imp.reload(fin_out)
 if "hul_in" in locals():
     imp.reload(hul_in)
+if "hul_out" in locals():
+    imp.reload(hul_out)
 if "img_in" in locals():
     imp.reload(img_in)
 if "prm_in" in locals():
@@ -63,13 +95,17 @@ if "w_in" in locals():
     imp.reload(w_in)
 if "w_out" in locals():
     imp.reload(w_out)
+if "rim_in" in locals():
+    imp.reload(rim_in)
+if "rim_out" in locals():
+    imp.reload(rim_out)
 
 
 # Makes common variables and classes directly accessible
 from .common import *
-from .props_mesh import *
-from .props_obj import *
-from .props_scene import *
+from .props.props_mesh import *
+from .props.props_obj import *
+from .props.props_scene import *
 from .texanim import *
 
 dprint("---\n\n\n\n")
@@ -77,7 +113,7 @@ dprint("---\n\n\n\n")
 bl_info = {
 "name": "Re-Volt",
 "author": "Marvin Thiel",
-"version": (18, 6, 19),
+"version": (18, 8, 1),
 "blender": (2, 79, 0),
 "location": "File > Import-Export",
 "description": "Import and export Re-Volt file formats.",
@@ -126,6 +162,7 @@ def register():
 
     bpy.types.INFO_MT_file_import.prepend(menu_func_import)
     bpy.types.INFO_MT_file_export.prepend(menu_func_export)
+    bpy.types.INFO_MT_add.append(menu_add.menu_func_add)
 
     bpy.app.handlers.scene_update_pre.append(edit_object_change_handler)
     # bpy.app.handlers.scene_update_post.append(edit_object_change_handler)
@@ -140,7 +177,7 @@ def unregister():
 
     bpy.types.INFO_MT_file_import.remove(menu_func_import)
     bpy.types.INFO_MT_file_export.remove(menu_func_export)
-
+    bpy.types.INFO_MT_add.remove(menu_add.menu_func_add)
 
 if __name__ == "__main__":
     register()