+++ io_revolt/ncp_out.py
@@ -41,7 +41,7 @@ def export_file(filepath, scene):
     # Collects objects for export
     objs = []
     if props.ncp_export_selected:
-        objs = [ob for ob in scene.objects if ob.select]
+        objs = [ob for ob in scene.objects if ob.select and not ob.revolt.ignore_ncp]
     else:
         for obj in scene.objects:
             conditions = (
@@ -50,7 +50,9 @@ def export_file(filepath, scene):
                 not obj.revolt.is_cube and
                 not obj.revolt.is_bcube and
                 not obj.revolt.is_bbox and
-                not obj.revolt.ignore_ncp
+                #not obj.revolt.is_instance and
+                not obj.revolt.ignore_ncp and
+                not obj.revolt.is_mirror_plane
             )
             if conditions:
                 objs.append(obj)
@@ -58,23 +60,54 @@ def export_file(filepath, scene):
     if objs == []:
         common.queue_error("exporting NCP", "No suitable objects in scene.")
         return
+    else:
+        dprint("Suitable objects: {}".format(", ".join([o.name for o in objs])))
 
     # Creates a mesh for all objects
-    transform = not (props.ncp_export_selected and len(objs) == 1)
-    bm = objects_to_bmesh(objs, transform)
+    transform = len(objs) != 1 or props.apply_translation
+    # bm = objects_to_bmesh(objs, transform) this breaks custom props
 
-    if props.triangulate_ngons:
-        num_ngons = triangulate_ngons(bm)
-        if scene.revolt.triangulate_ngons > 0:
-            print("Triangulated {} n-gons".format(num_ngons))
+    ncp = NCP()
+
+    # Adds all meshes to the ncp
+    for obj in objs:
+        dprint("Adding {} to ncp...".format(obj.name))
+        bm = bmesh.new()
+        bm.from_mesh(obj.data)
+
+        if props.triangulate_ngons:
+            num_ngons = triangulate_ngons(bm)
+            if scene.revolt.triangulate_ngons > 0:
+                print("Triangulated {} n-gons".format(num_ngons))
+
+        # Applies translation, rotation and scale
+        apply_trs(obj, bm, transform)
+
+        add_bm_to_ncp(bm, ncp)
+
+    # Sets length of polyhedron list
+    ncp.polyhedron_count = len(ncp.polyhedra)
+
+    # Creates a collision grid
+    if props.ncp_export_collgrid:
+        dprint("Exporting collision grid...")
+        ncp.generate_lookup_grid(grid_size=props.ncp_collgrid_size)
+
+    # Writes the NCP to file
+    with open(filepath, "wb") as f:
+        ncp.write(f)
+
+    # Frees the bmesh
+    bm.free()
+
+
+def add_bm_to_ncp(bm, ncp):
 
     # Material and type layers. The preview layer will be ignored.
     material_layer = (bm.faces.layers.int.get("Material") or
                       bm.faces.layers.int.new("Material"))
     type_layer = (bm.faces.layers.int.get("NCPType") or
                   bm.faces.layers.int.new("NCPType"))
-
-    ncp = NCP()
 
     for face in bm.faces:
         poly = Polyhedron()
@@ -90,7 +123,11 @@ def export_file(filepath, scene):
 
         # Sets polyhedron properties
         poly.material = face[material_layer]
-        print(face[material_layer])
+        if poly.material > 26:
+            print(face[material_layer])
+            queue_error("exporting to .ncp", "Invalid material")
+            if DEBUG:
+                return
         poly.type = face[type_layer]
 
         verts = face.verts
@@ -128,17 +165,3 @@ def export_file(filepath, scene):
         # Creates a bbox and adds the poly to the ncp
         poly.bbox = BoundingBox(data=rvbbox_from_verts(verts))
         ncp.polyhedra.append(poly)
-
-    # Sets length of polyhedron list
-    ncp.polyhedron_count = len(ncp.polyhedra)
-
-    # Creates a collision grid
-    if props.ncp_export_collgrid:
-        ncp.generate_lookup_grid(grid_size=props.ncp_collgrid_size)
-
-    # Writes the NCP to file
-    with open(filepath, "wb") as f:
-        ncp.write(f)
-
-    # Frees the bmesh
-    bm.free()